stock-watchlist/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
yfinance>=0.2.40
pandas>=2.2.2
pydantic[email]>=2.7.1
pydantic-settings>=2.6.0
python-multipart>=0.0.9
websockets>=12.0
httpx>=0.27.0
aiosqlite>=0.20.0
pandas-ta>=0.3.14b1
finnhub-python>=2.4.20
